Adjust tests to long values

+++ src/test/kotlin/ChunkDownloadTests.kt
@@ -13,11 +13,11 @@ class ChunkDownloadTests {
         val destination = Files.createTempFile("downloader-chunk", ".bin")
         try {
             runBlocking {
-                downloader.downloadChunk("${TestFixtures.serverUrl}/${TestFixtures.fileName}", destination, 0 to 512)
+                downloader.downloadChunk("${TestFixtures.serverUrl}/${TestFixtures.fileName}", destination, 0L to 512L)
             }
 
             val downloadedBytes = Files.readAllBytes(destination)
-            assertEquals(512, downloadedBytes.size)
+            assertEquals(512L, downloadedBytes.size.toLong())
             assertArrayEquals(TestFixtures.testFileData.copyOfRange(0, 512), downloadedBytes)
         } finally {
             Files.deleteIfExists(destination)
@@ -27,15 +27,15 @@ class ChunkDownloadTests {
     @Test
     fun testDownloadChunkWritesAtTheRequestedOffset() {
         val destination = Files.createTempFile("downloader-offset", ".bin")
-        Files.write(destination, ByteArray(TestFixtures.testFileSize))
+        Files.write(destination, ByteArray(TestFixtures.testFileSize.toInt()))
 
         try {
             runBlocking {
-                downloader.downloadChunk("${TestFixtures.serverUrl}/${TestFixtures.fileName}", destination, 512 to 1024)
+                downloader.downloadChunk("${TestFixtures.serverUrl}/${TestFixtures.fileName}", destination, 512L to 1024L)
             }
 
             val downloadedBytes = Files.readAllBytes(destination)
-            assertEquals(TestFixtures.testFileSize, downloadedBytes.size)
+            assertEquals(TestFixtures.testFileSize, downloadedBytes.size.toLong())
             assertTrue(downloadedBytes.copyOfRange(0, 512).all { it == 0.toByte() })
             assertArrayEquals(TestFixtures.testFileData.copyOfRange(512, 1024), downloadedBytes.copyOfRange(512, 1024))
         } finally {
